Huge refactor
The following changes were made:

 * Added `names` feature, which deserialises name fields.
 * Added `Checked` type, which is used for validating data during the
   deserialisation process. This is important since we want validation
   errors to occur during the validation process and not during
   deserialisation.
 * Implemented poor man's zero-copy deserialisation, whereby
   stringly typed constants are converted to checked enums. This reduces
   the size of JSON data structures and also reduces the number of allocations
   required for deserialisation.
 * Cow<'a, str> is now String, but always opt-in, requiring either the `names` or
   the `extras` feature to be enabled.

+++ src/json/root.rs
@@ -7,6 +7,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+use serde;
 use std;
 use std::fmt;
 use json::*;
@@ -31,7 +32,6 @@ pub struct Index<T>(u32, std::marker::PhantomData<T>);
 
 /// The root object of a glTF 2.0 asset.
 #[derive(Clone, Debug, Default, Deserialize, Validate)]
-#[serde(deny_unknown_fields)]
 pub struct Root {
     /// An array of accessors.
     #[serde(default)]
@@ -290,8 +290,8 @@ impl<T> Index<T> {
     }
 
     /// Returns the internal offset value.
-    pub fn value(&self) -> u32 {
-        self.0
+    pub fn value(&self) -> usize {
+        self.0 as usize
     }
 }
